fix: setup integration and system tests

+++ tests/integration/infrastructure/db/repositories/brand/dynamo-brand-repository.spec.ts
@@ -3,22 +3,46 @@ import { DocumentClient } from "aws-sdk/clients/dynamodb";
 import { DBIndexPrefixes } from "../../../../../../src/infrastructure/db/enums/db-index-prefixes";
 import { DynamoBrandRepository } from "../../../../../../src/infrastructure/db/repositories/brand/dynamo-brand-repository";
 import { IBrandRepository } from "../../../../../../src/infrastructure/db/repositories/brand/ibrand-repository";
-import { tableModel } from "../../../../../../src/infrastructure/db/settings/table-model";
 
 const dynamoDb = new DynamoDB({ endpoint: process.env.TABLE_ENDPOINT });
 
 const prepareEnvironment = async () => {
   try {
     const tables = await dynamoDb.listTables().promise();
-    const foundTable = tables.TableNames?.some((name: string) =>
-      [process.env.TABLE_NAME].includes(name)
+    const foundTable = tables.TableNames?.some(
+      (name: string) => [process.env.TABLE_NAME].includes(name)
     );
 
     if (foundTable) return;
 
-    await dynamoDb
-      .createTable(tableModel)
-      .promise();
+    await dynamoDb.createTable({
+      AttributeDefinitions: [
+        { AttributeName: "pk", AttributeType: "S" },
+        { AttributeName: "sk", AttributeType: "S" },
+        { AttributeName: "gsi1pk", AttributeType: "S" },
+        { AttributeName: "gsi1sk", AttributeType: "S" },
+      ],
+      KeySchema: [
+        { AttributeName: "pk", KeyType: "HASH" },
+        { AttributeName: "sk", KeyType: "RANGE" },
+      ],
+      GlobalSecondaryIndexes: [
+        {
+          IndexName: "gsi1",
+          KeySchema: [
+            { AttributeName: "gsi1pk", KeyType: "HASH" },
+            { AttributeName: "gsi1sk", KeyType: "RANGE" },
+          ],
+          ProvisionedThroughput: { ReadCapacityUnits: 10, WriteCapacityUnits: 10 },
+          Projection: {
+            ProjectionType: "ALL",
+          },
+        },
+      ],
+      ProvisionedThroughput: { ReadCapacityUnits: 10, WriteCapacityUnits: 10 },
+      TableName: process.env.TABLE_NAME,
+    }).promise();
+
   } catch (error) {
     console.log(error);
   }
@@ -31,6 +55,7 @@ const teardownEnvironment = async () => {
       [process.env.TABLE_NAME].includes(name)
     );
 
+    console.log('foundTable >> ', foundTable);
     if (!foundTable) return;
 
     await dynamoDb.deleteTable({ TableName: process.env.TABLE_NAME }).promise();
@@ -40,12 +65,10 @@ const teardownEnvironment = async () => {
 };
 
 const makeDynamoBrandRepository = (): IBrandRepository => {
-  const settings = {
+  return new DynamoBrandRepository({
     tableEndpoint: process.env.TABLE_ENDPOINT,
     tableName: process.env.TABLE_NAME,
-  };
-
-  return new DynamoBrandRepository(settings);
+  });
 };
 
 const defaultNewBrandData = { 
@@ -56,18 +79,23 @@ const defaultNewBrandData = {
 
 describe("Integration", () => {
   describe("Infastructure::DB::Repositories::Brand", () => {
-    beforeAll(teardownEnvironment);
     beforeEach(prepareEnvironment);
     afterEach(teardownEnvironment);
 
     describe("DynamoBrandRepository.store()", () => {
       it("Should throw an error if DynamoDB throws", async () => {
         // Given
-        // forcing an error because the table does not exists.
-        await dynamoDb
-          .deleteTable({ TableName: process.env.TABLE_NAME })
-          .promise();
         const brandRepository = makeDynamoBrandRepository();
+        // forcing an error because the table does not exists.
+        const tables = await dynamoDb.listTables().promise();
+        const foundTable = tables.TableNames?.some(
+          (name: string) => [process.env.TABLE_NAME].includes(name)
+        );
+        if(foundTable) {
+          await dynamoDb
+            .deleteTable({ TableName: process.env.TABLE_NAME })
+            .promise();
+        }
 
         // When
         const result = brandRepository.store(defaultNewBrandData);
